extract-layer/requirements.txt
google-cloud-storage
kafka-python==2.0.2
requests
python-gnupg 
fs
slack_sdk
sh